requirements.txt
flask==1.1.2
pyjwt==2.0.0
datetime
uuid
Flask-SQLAlchemy
